+++ packages/opds-web-client/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+### v0.2.9
+- Fixed function binding issue that caused a TypeError when resizing the window.
+- Added error handler in DataFetcher if an adapter is not configured.
+
 ### v0.2.8
 - Added Typedoc for code documentation of React components and related classes and functions.